Split large files into focused partials for easier navigation
Game.cs split into Game.{Camera,DirtyRects,GpuAuxBuilder,Input,RenderPipeline}.cs partials plus standalone DrawProfile.cs and PerfCaptureSession.cs.
Terrain.cs split into TerrainGrid.{Heat,Render,Materialization,Decoration}.cs partials.
Backend.Native.cs split into Backend.{TerrainAux,PostParams}.cs partials.
TankHeatBehaviorTests.cs split into TankHeatBehaviorTests + WorldThermalSimulationTests with shared TankHeatTestHelpers.
No behavior changes; all 21 heat/thermal tests pass.
