+++ src/Tunnerer.Desktop/Game.Input.cs
@@ -1,0 +1,58 @@
+namespace Tunnerer.Desktop;
+
+using Tunnerer.Core.Types;
+using Tunnerer.Desktop.Config;
+
+public partial class Game
+{
+    private DirectionF? ComputeAimDirection(IReadOnlyList<Core.Entities.Tank> tanks)
+    {
+        if (tanks.Count == 0) return null;
+        var tank = tanks[0];
+
+        if (!IsMouseInViewport(out float relX, out float relY))
+            return null;
+
+        int scale = DesktopScreenTweaks.PixelScale;
+        float aimWorldX = (_camPixelX + relX) / scale;
+        float aimWorldY = (_camPixelY + relY) / scale;
+
+        float dx = aimWorldX - tank.Position.X;
+        float dy = aimWorldY - tank.Position.Y;
+        float len = MathF.Sqrt(dx * dx + dy * dy);
+        if (len < 0.001f) return null;
+        return new DirectionF(dx / len, dy / len);
+    }
+
+    private bool IsMouseInViewport(out float relX, out float relY)
+    {
+        var (mx, my, _) = _renderer.GetMouseState();
+        (float x, float y, float w, float h) vp = _renderBackend.SupportsUi
+            ? _hud.ViewportRect
+            : (0f, 0f, _hiResSize.X, _hiResSize.Y);
+
+        relX = 0; relY = 0;
+        if (vp.w <= 0 || vp.h <= 0) return false;
+
+        relX = mx - vp.x;
+        relY = my - vp.y;
+        if (relX < 0 || relY < 0 || relX >= vp.w || relY >= vp.h)
+            return false;
+
+        if (_renderBackend.SupportsUi)
+        {
+            float sx = _hiResSize.X / Math.Max(1f, vp.w);
+            float sy = _hiResSize.Y / Math.Max(1f, vp.h);
+            relX *= sx;
+            relY *= sy;
+        }
+
+        return true;
+    }
+
+    private bool IsLeftMouseDown()
+    {
+        var (_, _, buttons) = _renderer.GetMouseState();
+        return (buttons & 1) != 0;
+    }
+}
