+++ src/Tunnerer.Desktop/Rendering/Dx11/Backend.Native.cs
@@ -176,75 +176,6 @@ public sealed unsafe partial class Backend
         _postRtv = postRtv;
     }
 
-    private void UpdateTerrainAuxTexture(byte[]? auxData, Size worldSize, Rect? dirtyRect)
-    {
-        int worldW = worldSize.X;
-        int worldH = worldSize.Y;
-        if (auxData == null || worldW <= 0 || worldH <= 0)
-            return;
-
-        bool sizeChanged = _terrainAuxTexture == null || _terrainAuxW != worldW || _terrainAuxH != worldH;
-        if (sizeChanged)
-        {
-            if (_terrainAuxSrv != null) { _terrainAuxSrv->Release(); _terrainAuxSrv = null; }
-            if (_terrainAuxTexture != null) { _terrainAuxTexture->Release(); _terrainAuxTexture = null; }
-
-            var desc = new Texture2DDesc
-            {
-                Width = (uint)worldW,
-                Height = (uint)worldH,
-                MipLevels = 1,
-                ArraySize = 1,
-                Format = Format.FormatR8G8B8A8Unorm,
-                SampleDesc = new SampleDesc(1, 0),
-                Usage = Usage.Default,
-                BindFlags = (uint)BindFlag.ShaderResource,
-                CPUAccessFlags = 0,
-                MiscFlags = 0,
-            };
-            fixed (byte* ptr = auxData)
-            {
-                var sub = new SubresourceData
-                {
-                    PSysMem = ptr,
-                    SysMemPitch = (uint)(worldW * 4),
-                    SysMemSlicePitch = 0,
-                };
-                ID3D11Texture2D* auxTex = null;
-                _device->CreateTexture2D(&desc, &sub, &auxTex);
-                _terrainAuxTexture = auxTex;
-            }
-            ID3D11ShaderResourceView* auxSrv = null;
-            _device->CreateShaderResourceView((ID3D11Resource*)_terrainAuxTexture, null, &auxSrv);
-            _terrainAuxSrv = auxSrv;
-            _terrainAuxW = worldW;
-            _terrainAuxH = worldH;
-            return;
-        }
-
-        if (dirtyRect is not Rect rect)
-            return;
-        RectMath.GetMinMaxInclusive(rect, out int minX, out int minY, out int maxX, out int maxY);
-        int rw = maxX - minX + 1;
-        int rh = maxY - minY + 1;
-        if (rw <= 0 || rh <= 0) return;
-        var box = new Box((uint)minX, (uint)minY, 0u, (uint)(maxX + 1), (uint)(maxY + 1), 1u);
-        int required = rw * rh * 4;
-        if (_terrainAuxUploadScratch == null || _terrainAuxUploadScratch.Length < required)
-            _terrainAuxUploadScratch = new byte[required];
-        byte[] scratch = _terrainAuxUploadScratch;
-        for (int y = 0; y < rh; y++)
-        {
-            int src = ((minY + y) * worldW + minX) * 4;
-            int dst = y * rw * 4;
-            Array.Copy(auxData, src, scratch, dst, rw * 4);
-        }
-        fixed (byte* ptr = scratch)
-        {
-            _context->UpdateSubresource((ID3D11Resource*)_terrainAuxTexture, 0, &box, ptr, (uint)(rw * 4), 0);
-        }
-    }
-
     private void RunPostProcessPass(in GamePixelsUpload upload)
     {
         if (_postRtv == null || _sceneSrv == null || _postPs == null || _postParamsBuffer == null)
@@ -294,103 +225,4 @@ public sealed unsafe partial class Backend
         _context->PSSetSamplers(0, 1, &sampler);
     }
 
-    private void UpdatePostParamsBuffer(in GamePixelsUpload upload)
-    {
-        if (_postParamsBuffer == null)
-            return;
-        PostParamsCBuffer cbData = default;
-        cbData.TexelSizeX = 1f / Math.Max(1, upload.View.ViewSize.X);
-        cbData.TexelSizeY = 1f / Math.Max(1, upload.View.ViewSize.Y);
-        cbData.PixelScale = upload.View.PixelScale;
-        cbData.Time = (float)ImGui.GetTime();
-        cbData.WorldSizeX = upload.View.WorldSize.X;
-        cbData.WorldSizeY = upload.View.WorldSize.Y;
-        cbData.CameraPixelsX = upload.View.CameraPixels.X;
-        cbData.CameraPixelsY = upload.View.CameraPixels.Y;
-        cbData.ViewSizeX = upload.View.ViewSize.X;
-        cbData.ViewSizeY = upload.View.ViewSize.Y;
-        cbData.UseTerrainAux = upload.TerrainAux != null ? 1f : 0f;
-        cbData.BloomThreshold = DesktopScreenTweaks.PostBloomThreshold;
-        cbData.BloomStrength = DesktopScreenTweaks.PostBloomStrength;
-        cbData.BloomWeightCenter = DesktopScreenTweaks.PostBloomWeightCenter;
-        cbData.BloomWeightAxis = DesktopScreenTweaks.PostBloomWeightAxis;
-        cbData.BloomWeightDiagonal = DesktopScreenTweaks.PostBloomWeightDiagonal;
-        cbData.VignetteStrength = DesktopScreenTweaks.PostVignetteStrength;
-        cbData.EdgeLightStrength = DesktopScreenTweaks.PostTerrainEdgeLightStrength;
-        cbData.EdgeLightBias = DesktopScreenTweaks.PostTerrainEdgeLightBias;
-        cbData.TankHeatGlowR = DesktopScreenTweaks.PostTankHeatGlowR;
-        cbData.TankHeatGlowG = DesktopScreenTweaks.PostTankHeatGlowG;
-        cbData.TankHeatGlowB = DesktopScreenTweaks.PostTankHeatGlowB;
-        cbData.TankHeatGlowA = 0f;
-        cbData.TerrainHeatGlowR = DesktopScreenTweaks.PostTerrainHeatGlowR;
-        cbData.TerrainHeatGlowG = DesktopScreenTweaks.PostTerrainHeatGlowG;
-        cbData.TerrainHeatGlowB = DesktopScreenTweaks.PostTerrainHeatGlowB;
-        cbData.TerrainHeatThreshold = DesktopScreenTweaks.PostTerrainHeatThreshold;
-        cbData.TerrainMaskEdgeStrength = DesktopScreenTweaks.PostTerrainMaskEdgeStrength;
-        cbData.TerrainMaskCaveDarken = DesktopScreenTweaks.PostTerrainMaskCaveDarken;
-        cbData.TerrainMaskSolidLift = DesktopScreenTweaks.PostTerrainMaskSolidLift;
-        cbData.TerrainMaskOutlineDarken = DesktopScreenTweaks.PostTerrainMaskOutlineDarken;
-        cbData.TerrainMaskRimLift = DesktopScreenTweaks.PostTerrainMaskRimLift;
-        cbData.TerrainMaskBoundaryScale = DesktopScreenTweaks.PostTerrainMaskBoundaryScale;
-        cbData.VignetteInnerRadius = DesktopScreenTweaks.PostVignetteInnerRadius;
-        cbData.VignetteOuterRadius = DesktopScreenTweaks.PostVignetteOuterRadius;
-        cbData.Quality = (float)upload.Quality;
-        cbData.MaterialEnergyR = DesktopScreenTweaks.PostMaterialEmissiveEnergyR;
-        cbData.MaterialEnergyG = DesktopScreenTweaks.PostMaterialEmissiveEnergyG;
-        cbData.MaterialEnergyB = DesktopScreenTweaks.PostMaterialEmissiveEnergyB;
-        cbData.MaterialEnergyStrength = DesktopScreenTweaks.PostMaterialEmissiveEnergyStrength;
-        cbData.MaterialScorchedR = DesktopScreenTweaks.PostMaterialEmissiveScorchedR;
-        cbData.MaterialScorchedG = DesktopScreenTweaks.PostMaterialEmissiveScorchedG;
-        cbData.MaterialScorchedB = DesktopScreenTweaks.PostMaterialEmissiveScorchedB;
-        cbData.MaterialScorchedStrength = DesktopScreenTweaks.PostMaterialEmissiveScorchedStrength;
-        cbData.MaterialPulseFreq = DesktopScreenTweaks.PostMaterialEmissivePulseFreq;
-        cbData.MaterialPulseMin = DesktopScreenTweaks.PostMaterialEmissivePulseMin;
-        cbData.MaterialPulseRange = DesktopScreenTweaks.PostMaterialEmissivePulseRange;
-        cbData.MaterialPulsePad = 0f;
-        cbData.NativeEdgeSoftness = DesktopScreenTweaks.NativeContinuousEdgeSoftness;
-        cbData.NativeBoundaryBlend = DesktopScreenTweaks.NativeContinuousBoundaryBlend;
-        cbData.NativeSampleFactor = upload.NativeSampleCount <= 1 ? 0f : upload.NativeSampleCount <= 2 ? 0.5f : 1f;
-        cbData.NativePad = 0f;
-
-        float lx = DesktopScreenTweaks.LightDirX, ly = DesktopScreenTweaks.LightDirY, lz = DesktopScreenTweaks.LightDirZ;
-        float lLen = MathF.Sqrt(lx * lx + ly * ly + lz * lz);
-        if (lLen > 0.001f) { lx /= lLen; ly /= lLen; lz /= lLen; }
-        cbData.LightDirX = lx;
-        cbData.LightDirY = ly;
-        cbData.LightDirZ = lz;
-        cbData.LightNormalStrength = DesktopScreenTweaks.LightNormalStrength;
-        float hx = lx, hy = ly, hz = lz + 1f;
-        float hLen = MathF.Sqrt(hx * hx + hy * hy + hz * hz);
-        if (hLen > 0.001f) { hx /= hLen; hy /= hLen; hz /= hLen; }
-        cbData.HalfVecX = hx;
-        cbData.HalfVecY = hy;
-        cbData.HalfVecZ = hz;
-        cbData.LightMicroNormalStrength = DesktopScreenTweaks.LightMicroNormalStrength;
-        cbData.LightAmbient = DesktopScreenTweaks.LightAmbient;
-        cbData.LightDiffuseWeight = DesktopScreenTweaks.LightDiffuseWeight;
-        cbData.LightShininess = DesktopScreenTweaks.LightShininess;
-        cbData.LightSpecularIntensity = DesktopScreenTweaks.LightSpecularIntensity;
-
-        int glowCount = Math.Clamp(upload.TankHeatGlowCount, 0, 8);
-        cbData.TankGlowCount = glowCount;
-        if (upload.TankHeatGlowData != null)
-        {
-            for (int i = 0; i < glowCount; i++)
-            {
-                int src = i * 4;
-                int dst = i * 4;
-                cbData.TankGlow[dst + 0] = upload.TankHeatGlowData[src + 0];
-                cbData.TankGlow[dst + 1] = upload.TankHeatGlowData[src + 1];
-                cbData.TankGlow[dst + 2] = upload.TankHeatGlowData[src + 2];
-                cbData.TankGlow[dst + 3] = upload.TankHeatGlowData[src + 3];
-            }
-        }
-
-        MappedSubresource mapped;
-        if (_context->Map((ID3D11Resource*)_postParamsBuffer, 0, Silk.NET.Direct3D11.Map.WriteDiscard, 0, &mapped) >= 0)
-        {
-            *(PostParamsCBuffer*)mapped.PData = cbData;
-            _context->Unmap((ID3D11Resource*)_postParamsBuffer, 0);
-        }
-    }
 }
